+++ index.php
@@ -12,14 +12,19 @@ $task->setContractor($contractor);
 
 assert($task->getStatus() === TaskStatus::NEW,
     'A new task has the status "NEW"');
-assert($task->getActionFor($client) === TaskAction::CANCEL,
+
+assert($task->getActionFor($client)->getInternalName() === TaskAction::CANCEL,
     'A client has the action to "CANCEL" a new task');
-assert($task->getActionFor($contractor) === TaskAction::START,
+
+assert($task->getActionFor($contractor)->getInternalName() === TaskAction::START,
     'A contractor has the action to "START" a new task');
-assert($task->getActionFor($randomUser) === '',
+
+assert($task->getActionFor($randomUser) === null,
     'A random user doesn\'t have any actions');
+
 assert($task->getNextStatus($task->getActionFor($client), $client) === TaskStatus::CANCELED,
     'The next task\'s status for a client action is "CANCELED"');
+
 assert($task->getNextStatus($task->getActionFor($contractor), $contractor) === TaskStatus::PENDING,
     'The next task\'s status for a contractor action is "PENDING"');
