+++ src/models/CancelAction.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace TaskForce\models;
+
+class CancelAction extends AbstractAction
+{
+    protected $internalName = TaskAction::CANCEL;
+    protected $externalName = 'Отменить';
+
+    public function isAllowed(User $user, Task $task)
+    {
+        return $user->id === $task->client->id;
+    }
+}
+
