+++ src/models/Task.php
@@ -2,46 +2,6 @@
 
 namespace TaskForce\models;
 
-abstract class AbstractAction
-{
-    private $internalName;
-    private $externalName;
-
-    public function getInternalName()
-    {
-        return $this->internalName;
-    }
-
-    public function getExternalName()
-    {
-        return $this->externalName;
-    };
-
-    abstract protected function isAllowed(User $user, Task $task);
-}
-
-class StartAction extends AbstractAction
-{
-    private $internalName = TaskAction::START;
-    private $externalName = 'Откликнутся';
-
-    public function isAllowed(User $user, Task $task)
-    {
-        return $user->id === $task->contractor->id;
-    }
-}
-
-class CancelAction extends AbstractAction
-{
-    private $internalName = TaskAction::CANCEL;
-    private $externalName = 'Отменить';
-
-    public function isAllowed(User $user, Task $task)
-    {
-        return $user->id === $task->client->id;
-    }
-}
-
 class Task
 {
     public $id;
@@ -99,7 +59,7 @@ class Task
         $this->contractor = $user;
     }
 
-    public function getActionFor(User $user): ?string
+    public function getActionFor(User $user): ?AbstractAction
     {
         if ($user->id === $this->client->id) {
             return $this->actions[$this->status][UserRole::CLIENT];
@@ -116,12 +76,12 @@ class Task
         return $this->status;
     }
 
-    public function getNextStatus(string $action, User $user): ?string
+    public function getNextStatus(AbstractAction $action, User $user): ?string
     {
-        if ($this->getActionFor($user) !== $action) {
+        if ($this->getActionFor($user)->getInternalName() !== $action->getInternalName()) {
             return null;
         }
-        return $this->transitions[$action];
+        return $this->transitions[$action->getInternalName()];
     }
 
     public function setNextStatus(string $action, User $user)
